refactor: 메뉴 도메인 신규생성
