+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -2,9 +2,9 @@ package kitchenpos.menus.application;
 
 import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuGroup;
-import kitchenpos.menus.domain.MenuGroupRepository;
 import kitchenpos.menus.domain.MenuProduct;
-import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.menus.domain.tobe.MenuGroupRepository;
+import kitchenpos.menus.domain.tobe.MenuRepository;
 import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.tobe.BanWordFilter;
 import kitchenpos.products.domain.tobe.ProductRepository;
