+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -2,7 +2,7 @@ package kitchenpos.products.application;
 
 import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuProduct;
-import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.menus.domain.tobe.MenuRepository;
 import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.tobe.BanWordFilter;
 import kitchenpos.products.domain.tobe.ProductRepository;
